requirements.txt
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
blinker==1.5
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pycountry==22.3.5
PyMySQL==1.0.2
pytest==7.2.2
pytest-order==1.1.0
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4
tomli==2.0.1
typish==1.9.3
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
